pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "martin-config"
version = "0.0.1"
authors = [
    { name="Ioan Ferencik", email="ioan.ferencik@undp.org" },
    { name="Jin Igarashi", email="jin.igarashi@undp.org" }
]
description = "Generate configuration files for martin vector tile server"
readme = "README.md"

requires-python = ">=3.6"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "asyncpg", "pyyaml", "dotenv"
]
[project.scripts]
    martincfg = "martin_config:martincfg"

[project.optional-dependencies]
azure = ["azure-storage-file-share"]

[tool.setuptools.packages.find]
where = ["src"]



